# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMMAS API",
        "description": "Sistem Informasi Manajemen Magang Siswa (SIMMAS). REST API dengan autentikasi Sanctum. Token didapat dari endpoint `POST /api/login`. Semua endpoint selain `login` dan `public/*` memerlukan header `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server backend SIMMAS (Laravel)"
        }
    ],
    "paths": {
...
        "/api/siswa/pengajuan": {
            "get": {
                "tags": [
                    "Siswa"
                ],
                "summary": "Daftar pengajuan magang siswa",
                "description": "Riwayat pengajuan magang milik siswa yang sedang login.",
                "operationId": "6162e48c08275a1138c8e6292a1cecc0",
                "responses": {
                    "200": {
                        "description": "Daftar pengajuan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PengajuanMagang"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Siswa"
                ],
                "summary": "Buat pengajuan magang",
                "description": "Mengirim pengajuan magang ke DUDI. Kuota DUDI harus tersedia. Status siswa otomatis menjadi pengajuan.",
                "operationId": "0dc1de09f1c11d8a262efeff4014d944",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tempat_magang_id",
                                    "posisi_diminati",
                                    "tanggal_mulai_usulan",
                                    "tanggal_selesai_usulan"
                                ],
                                "properties": {
                                    "tempat_magang_id": {
                                        "type": "integer"
                                    },
                                    "posisi_diminati": {
                                        "type": "string",
                                        "example": "Programmer"
                                    },
                                    "tanggal_mulai_usulan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tanggal_selesai_usulan": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan berhasil dikirim.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PengajuanMagang"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau kuota DUDI habis."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PengajuanMagang": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "siswa_id": {
                        "type": "integer"
                    },
                    "tempat_magang_id": {
                        "type": "integer"
                    },
                    "posisi_diminati": {
                        "type": "string"
                    },
                    "tanggal_mulai_usulan": {
                        "type": "string",
                        "format": "date"
                    },
                    "tanggal_selesai_usulan": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "menunggu",
                            "disetujui",
                            "ditolak"
                        ]
                    },
                    "catatan_penolakan": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer <token> (didapat dari endpoint POST /api/login).",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}